+++ consensus/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Switch transaction validation call sites to use the explicit
+  `LedgerTransaction` ledger wrapper.
+
 ## [1.6.0] - 2026-02-27
 
 ### Changed